+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\BankApp",
+    "\\BankApp\\BankApp",
+    "\\BankApp\\BankApp\\Client"
+  ],
+  "SelectedNode": "\\BankApp\\BankApp\\BaseUser.h",
+  "PreviewInSolutionExplorer": false
+}
